rename "contact card" -> "public profile"
Zheming Lin pointed out there is already a standard for contact cards
called "vcard". However, this standard is actually not about the private
information that is usually shared inside of a contact card. It is for
the public profile information that anyone can access with no
authentication. It is specifically for name & avatar, and possibly other
public information like website. It is not for personal shipping address
or anything like that. Thus I am renaming it to "public profile" to
reflect the proper nature of the standard.